clippy module inception: move kernel::kernel into kernel
